feat: Add `help` command to interpreter for accessing language documentation
This commit adds a new `help` command to the interpreter, allowing users
to access documentation for the babyjs language. The command provides a
detailed overview of the language features and syntax, along with
examples of various programming constructs. The `help` command is
implemented as a native function in the `Interpreter` class and is
available for use in the REPL. This addition enhances the user
experience by providing a convenient way to access language
documentation directly from the interpreter.