feat: Add support for replacing output in interpreter interface
This change modifies the `useStd` hook in the `interface.tsx` file to
introduce the ability to replace the output in the interpreter
interface. Previously, the output was only appended. Now, by passing the
`printType` parameter as "replace" to `useStd`, the output is cleared
before new output is displayed.

This change improves the functionality of the interpreter interface by
providing more control over how the output is displayed, especially in
scenarios where replacing the output is desired instead of continuously
appending to it.